+++ Core/Src/pid.c
@@ -1,6 +1,6 @@
 #include "pid.h"
 
-void PID_init(PID_t *pid, double Kp, double Ki, double Kd, double Ts, double minOutput, double maxOutput)
+void PID_init(PID_t *pid, float Kp, float Ki, float Kd, float Ts, float minOutput, float maxOutput)
 {
 	pid->Kd = Kd;
 	pid->Ki = Ki;
@@ -12,18 +12,18 @@ void PID_init(PID_t *pid, double Kp, double Ki, double Kd, double Ts, double min
 	pid->sumError = 0;
 }
 
-double PID_calculate(PID_t *pid, int setpoint, int feedback)
+float PID_calculate(PID_t *pid, int setpoint, int feedback)
 {
 	float output;
 	float error = feedback - setpoint;
 
 	output = pid->Kp * error;
-    output += pid->Ki * pid->sumError * pid->Ts;
-    output += pid->Kd * (error - pid->previousError) / pid->Ts;
+	output += pid->Ki * pid->sumError * pid->Ts;
+	output += pid->Kd * (error - pid->previousError) / pid->Ts;
 
-    pid->previousError = error;
-    output = output > pid->maxOutput ? pid->maxOutput : output;
-    output = output < pid->minOutput ? pid->minOutput : output;
+	pid->previousError = error;
+	output = output > pid->maxOutput ? pid->maxOutput : output;
+	output = output < pid->minOutput ? pid->minOutput : output;
 
-    return output;
+	return output;
 }
